Update sample html file with zlib compressed file
The full sample html files exceeded the 50 MB limit for github.
Making a .zip version of the files avoided the limit but required
the user to download the .zip file, unzip it, and then load it
in the browswer.
Now I compress the string pool and chart_data with zlib when the
standalone html files are created. And then the data is
decompressed using a javascript implementation of zlib.
This reduces the size of a larger file from 80 MBs to 11 MBs.