Add more cpu_diagram info
Modified spin.cpp to be able to run multiple workloads in the same
run. SO I created a run with a memory bw test, an L3 bw test, an
L2 test and a spin (just an add loop) test.
Then I used the 'phase' info to zoom to each test and created a
PNG of each phases' cpu_diagram and table.
Added each PNG to the README.md.